Writing down next steps for LBVH implementation
Things to follow-up on the plane (probably can't do the complete sorting algorithm in the air, but I can do a lot of the rest).

If not the plane, then whenever I have enough headspace next ^_^'. Might be my trip to Guangzhou for Wenlin, idk.

A merge sort variation might be best, seems my thinking was wondering that way already - biggest conceptual difference is that in my notes here I only actually sort the pairs in the first pass, whereas the the merge-sort algorithm (correctly) sorts at every level (so it sorts within the pairs, but it also sorts within the 4-wide clumps, eight-wide clumps, etc)

https://moderngpu.github.io/mergesort.html

Radix sort seems faster, but also a much more involved implementation; might start with merge sort and change if performance is an issue.

https://gpuopen.com/download/publications/Introduction_to_GPU_Radix_Sort.pdf